remove any references to repository tiers
In practice, repository tiers did not provide additional context on
how or how well a given repository was maintained.  Given that we have
not found any precedent for these tiers in the broader Open Source
world, we are removing the tiers and relying on the license in each
repository to establish what, if any, maintenance expectations are
valid for that repository.